Added Profile features
