+++ TrainRec/trainrec_exercises.py
@@ -1,11 +1,55 @@
-from flask import Flask, jsonify, request
+from flask import Flask, jsonify, request, g  
 from flask_cors import CORS
 import uuid
+import sqlite3                                
+from datetime import datetime, timedelta
+from functools import wraps
+from werkzeug.security import generate_password_hash, check_password_hash
+import jwt
 
 app = Flask(__name__)
 
 # This allows your team's React app (port 5173) to access this API (port 5000)
 CORS(app)
+
+
+DB_PATH = "trainrec.db"
+JWT_SECRET = "replace-with-a-very-secret-key"
+JWT_ALGORITHM = "HS256"
+
+def get_db():
+    db = getattr(g, "_database", None)
+    if db is None:
+        db = g._database = sqlite3.connect(DB_PATH)
+        db.row_factory = sqlite3.Row
+    return db
+
+@app.teardown_appcontext
+def close_connection(exception):
+    db = getattr(g, "_database", None)
+    if db is not None:
+        db.close()
+
+def init_db():
+    conn = sqlite3.connect(DB_PATH)
+    c = conn.cursor()
+    c.execute("CREATE TABLE IF NOT EXISTS users (id INTEGER PRIMARY KEY AUTOINCREMENT, email TEXT UNIQUE, name TEXT, nickname TEXT, password_hash TEXT, sex TEXT, birthdate TEXT, height_in REAL, weight_lb REAL, fitness_activity TEXT, created_at TEXT)")
+    c.execute("CREATE TABLE IF NOT EXISTS goals (id INTEGER PRIMARY KEY AUTOINCREMENT, user_id INTEGER, target_weight_lb REAL, weekly_weight_change_lb REAL, activity_level TEXT, daily_calorie_target REAL, updated_at TEXT)")
+    conn.commit()
+    conn.close()
+
+def auth_required(fn):
+    @wraps(fn)
+    def wrapper(*args, **kwargs):
+        auth = request.headers.get("Authorization", "")
+        if not auth.startswith("Bearer "): return jsonify({"error": "Unauthorized"}), 401
+        try:
+            token = auth.split(" ")[1]
+            data = jwt.decode(token, JWT_SECRET, algorithms=[JWT_ALGORITHM])
+            request.user_id = data.get("user_id")
+        except: return jsonify({"error": "Invalid token"}), 401
+        return fn(*args, **kwargs)
+    return wrapper
 
 # ---------------------------------------------------------
 # EXERCISE LIBRARY DATA
@@ -45,6 +89,104 @@ EXERCISES = [
 # ---------------------------------------------------------
 # API ROUTES
 # ---------------------------------------------------------
+
+@app.route("/register", methods=["POST"])
+def register():
+    payload = request.get_json()
+    pw_hash = generate_password_hash(payload["password"])
+    db = get_db()
+    try:
+        db.execute("""INSERT INTO users (email, name, nickname, password_hash, sex, birthdate, height_in, weight_lb, created_at) 
+                   VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?)""",
+                   (payload["email"].lower(), payload["name"], payload.get("nickname"), pw_hash, 
+                    payload["sex"], payload["birthdate"], payload["height_in"], payload["weight_lb"], datetime.utcnow().isoformat()))
+        db.commit()
+        return jsonify({"message": "Account Created"}), 201
+    except:
+        return jsonify({"error": "User already exists"}), 400
+
+@app.route("/login", methods=["POST"])
+def login():
+    payload = request.get_json() or {}
+    email = payload.get("email", "").lower()
+    password = payload.get("password", "")
+
+    db = get_db()
+    # Find the user by email
+    user = db.execute("SELECT id, password_hash FROM users WHERE email = ?", (email,)).fetchone()
+
+    if user and check_password_hash(user["password_hash"], password):
+        # Create the token (valid for 24 hours)
+        token = jwt.encode(
+            {"user_id": user["id"], "exp": datetime.utcnow() + timedelta(seconds=86400)}, 
+            JWT_SECRET, 
+            algorithm=JWT_ALGORITHM
+        )
+        return jsonify({
+            "message": "Logged in successfully",
+            "token": token
+        }), 200
+    
+    return jsonify({"error": "Invalid email or password"}), 401
+
+@app.route("/reset_password", methods=["POST"])
+@auth_required
+def reset_password():
+    new_pw = request.get_json().get("new_password")
+    db = get_db()
+    db.execute("UPDATE users SET password_hash = ? WHERE id = ?", (generate_password_hash(new_pw), request.user_id))
+    db.commit()
+    return jsonify({"message": "Password updated"})
+
+@app.route("/me/update_profile", methods=["POST"])
+@auth_required
+def update_profile():
+    payload = request.get_json()
+    db = get_db()
+    # Dynamic update for weight, height, nickname, and fitness activity
+    for key in ["weight_lb", "height_in", "nickname", "fitness_activity"]:
+        if key in payload:
+            db.execute(f"UPDATE users SET {key} = ? WHERE id = ?", (payload[key], request.user_id))
+    db.commit()
+    return jsonify({"message": "Profile updated"})
+
+@app.route("/goals", methods=["POST"])
+@auth_required
+def set_goals():
+    payload = request.get_json()
+    db = get_db()
+    user = db.execute("SELECT * FROM users WHERE id = ?", (request.user_id,)).fetchone()
+    
+    calories = calculate_daily_calories(user["sex"], user["weight_lb"], user["height_in"], 
+                                       user["birthdate"], payload["activity_level"], payload["weekly_weight_change_lb"])
+
+    db.execute("""INSERT INTO goals (user_id, target_weight_lb, weekly_weight_change_lb, activity_level, daily_calorie_target, updated_at)
+                  VALUES (?, ?, ?, ?, ?, ?)""", 
+               (request.user_id, payload["target_weight_lb"], payload["weekly_weight_change_lb"], 
+                payload["activity_level"], calories, datetime.utcnow().isoformat()))
+    db.commit()
+    return jsonify({"message": "Goal set", "daily_calories": calories})
+
+@app.route("/dashboard", methods=["GET"])
+@auth_required
+def dashboard():
+    db = get_db()
+    user = db.execute("SELECT * FROM users WHERE id = ?", (request.user_id,)).fetchone()
+    
+    # Auto-age for dashboard
+    birth = datetime.fromisoformat(user["birthdate"])
+    age = datetime.utcnow().year - birth.year
+    
+    return jsonify({
+        "display_name": user["nickname"] or user["name"],
+        "age": age,
+        "stats": {
+            "weight": f"{user['weight_lb']} lbs",
+            "height": f"{user['height_in']} inches",
+            "activity": user["fitness_activity"]
+        }
+    })
+
 # Temporary storage for active sessions
 active_sessions = {}
 
@@ -91,5 +233,6 @@ def health_check():
     return jsonify({"status": "online", "version": "2.0.0"})
 
 if __name__ == "__main__":
+    init_db()
     # Standard port 5000 is used so Vite proxy works correctly.
     app.run(debug=True, port=5000)
